refactor: update package to use new nameclature
