+++ jnosql-cassandra/src/main/java/org/eclipse/jnosql/databases/cassandra/mapping/CassandraUDTType.java
@@ -16,7 +16,7 @@ package org.eclipse.jnosql.databases.cassandra.mapping;
 
 import org.eclipse.jnosql.communication.column.Column;
 import org.eclipse.jnosql.databases.cassandra.communication.UDT;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.column.ColumnEntityConverter;
 import org.eclipse.jnosql.mapping.column.ColumnFieldValue;
 import org.eclipse.jnosql.mapping.metadata.FieldMetadata;
