+++ jnosql-cassandra/src/test/java/org/eclipse/jnosql/databases/cassandra/integration/TemplateIntegrationTest.java
@@ -22,7 +22,7 @@ import org.eclipse.jnosql.communication.driver.ConfigurationReader;
 import org.eclipse.jnosql.databases.cassandra.communication.CassandraConfigurations;
 import org.eclipse.jnosql.databases.cassandra.communication.ColumnDatabase;
 import org.eclipse.jnosql.mapping.Convert;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.column.ColumnEntityConverter;
 import org.eclipse.jnosql.mapping.column.spi.ColumnExtension;
 import org.eclipse.jnosql.mapping.config.MappingConfigurations;
