+++ jnosql-couchbase/src/main/java/org/eclipse/jnosql/databases/couchbase/mapping/CouchbaseDocumentRepositoryProxy.java
@@ -17,19 +17,19 @@ package org.eclipse.jnosql.databases.couchbase.mapping;
 
 import com.couchbase.client.java.json.JsonObject;
 import jakarta.data.repository.PageableRepository;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.document.JNoSQLDocumentTemplate;
 import org.eclipse.jnosql.mapping.document.query.AbstractDocumentRepositoryProxy;
 import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
 import org.eclipse.jnosql.mapping.metadata.EntityMetadata;
-import org.eclipse.jnosql.mapping.repository.DynamicReturn;
+import org.eclipse.jnosql.mapping.core.repository.DynamicReturn;
 
 import java.lang.reflect.Method;
 import java.lang.reflect.ParameterizedType;
 import java.util.Objects;
 import java.util.stream.Stream;
 
-import static org.eclipse.jnosql.mapping.repository.DynamicReturn.toSingleResult;
+import static org.eclipse.jnosql.mapping.core.repository.DynamicReturn.toSingleResult;
 
 
 class CouchbaseDocumentRepositoryProxy<T>extends AbstractDocumentRepositoryProxy<T> {
