+++ jnosql-couchbase/src/test/java/org/eclipse/jnosql/databases/couchbase/mapping/DefaultCouchbaseTemplateTest.java
@@ -20,7 +20,7 @@ import jakarta.inject.Inject;
 import org.eclipse.jnosql.communication.document.Document;
 import org.eclipse.jnosql.communication.document.DocumentEntity;
 import org.eclipse.jnosql.databases.couchbase.communication.CouchbaseDocumentManager;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.document.DocumentEntityConverter;
 import org.eclipse.jnosql.mapping.document.DocumentEventPersistManager;
 import org.eclipse.jnosql.mapping.document.spi.DocumentExtension;
